Making the state machine the actor, but hitting the wall that the previous design tried to avoid
It would be nice if the (actor) self reference was provided by the protocol extension.
That is, since we need the actor self reference needs the self reference of the concrete struct conforming to Actor we're caught in a catch-22
Is there no way to tell the compiler that things are going to be just fine?